+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [6.2.0] - 2022-01-28
+### Added
+- Uniswap OTA updates
+- Uniswap MongoDB price management 
+- Swagger documentation on /docs.
+- Added RPC Selection before minting
 
 ## [6.1.1] - 2022-01-27
 ### Added